fix(substrate): single-pass schema validation + i32 shape_version
Two review findings:
- The commit path compiled and validated the JSON Schema twice (once
  for the verdict decision, once for the detail record). Now validates
  once and reuses the owned error strings.
- ShapeDeclared.shape_version was i64 while the SQL column, readback,
  and wire types all use i32. Aligned to i32; schema fixture regenerated
  (int64 → int32).

+++ crates/temper-substrate/src/payloads.rs
@@ -726,7 +726,7 @@ pub struct ShapeDeclared {
     /// The chain depth of the assert/fold lineage — 1 for the first declaration, N for the Nth
     /// amendment. A verdict recorded against a folded version stops matching, and the artifact
     /// reads as unchecked until reconciled.
-    pub shape_version: i64,
+    pub shape_version: i32,
 }
 
 /// Set a **single-valued** property (WS6 4c): the projection folds prior active rows for
